requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.5
python-multipart==0.0.6
pusher==3.3.2
websockets==11.0.3 